Remove most calls to `Artifice.deactivate`
By reorganizing code so that we don't redefine `Net:HTTP`, then undefine
the redefinition, then define it again.
