+++ bundler/spec/support/artifice/compact_index_api_missing.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
-require_relative "compact_index"
-
-Artifice.deactivate
+require_relative "helpers/compact_index"
 
 class CompactIndexApiMissing < CompactIndexAPI
   get "/fetch/actual/gem/:id" do
     halt 404
   end
 end
+
+require_relative "helpers/artifice"
 
 Artifice.activate_with(CompactIndexApiMissing)
