+++ bundler/spec/support/artifice/compact_index_checksum_mismatch.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "compact_index"
-
-Artifice.deactivate
+require_relative "helpers/compact_index"
 
 class CompactIndexChecksumMismatch < CompactIndexAPI
   get "/versions" do
@@ -12,5 +10,7 @@ class CompactIndexChecksumMismatch < CompactIndexAPI
     body ""
   end
 end
+
+require_relative "helpers/artifice"
 
 Artifice.activate_with(CompactIndexChecksumMismatch)
