+++ bundler/spec/support/artifice/compact_index_concurrent_download.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "compact_index"
-
-Artifice.deactivate
+require_relative "helpers/compact_index"
 
 class CompactIndexConcurrentDownload < CompactIndexAPI
   get "/versions" do
@@ -28,5 +26,7 @@ class CompactIndexConcurrentDownload < CompactIndexAPI
     end
   end
 end
+
+require_relative "helpers/artifice"
 
 Artifice.activate_with(CompactIndexConcurrentDownload)
