+++ bundler/spec/support/artifice/compact_index_creds_diff_host.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "compact_index"
-
-Artifice.deactivate
+require_relative "helpers/compact_index"
 
 class CompactIndexCredsDiffHost < CompactIndexAPI
   helpers do
@@ -35,5 +33,7 @@ class CompactIndexCredsDiffHost < CompactIndexAPI
     end
   end
 end
+
+require_relative "helpers/artifice"
 
 Artifice.activate_with(CompactIndexCredsDiffHost)
