+++ bundler/spec/support/artifice/compact_index_forbidden.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
-require_relative "compact_index"
-
-Artifice.deactivate
+require_relative "helpers/compact_index"
 
 class CompactIndexForbidden < CompactIndexAPI
   get "/versions" do
     halt 403
   end
 end
+
+require_relative "helpers/artifice"
 
 Artifice.activate_with(CompactIndexForbidden)
