+++ bundler/spec/support/artifice/compact_index_no_gem.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
-require_relative "compact_index"
-
-Artifice.deactivate
+require_relative "helpers/compact_index"
 
 class CompactIndexNoGem < CompactIndexAPI
   get "/gems/:id" do
     halt 500
   end
 end
+
+require_relative "helpers/artifice"
 
 Artifice.activate_with(CompactIndexNoGem)
