+++ bundler/spec/support/artifice/compact_index_partial_update.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "compact_index"
-
-Artifice.deactivate
+require_relative "helpers/compact_index"
 
 class CompactIndexPartialUpdate < CompactIndexAPI
   # Stub the server to never return 304s. This simulates the behaviour of
@@ -34,5 +32,7 @@ class CompactIndexPartialUpdate < CompactIndexAPI
     end
   end
 end
+
+require_relative "helpers/artifice"
 
 Artifice.activate_with(CompactIndexPartialUpdate)
