+++ bundler/spec/support/artifice/compact_index_strict_basic_authentication.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "compact_index"
-
-Artifice.deactivate
+require_relative "helpers/compact_index"
 
 class CompactIndexStrictBasicAuthentication < CompactIndexAPI
   before do
@@ -16,5 +14,7 @@ class CompactIndexStrictBasicAuthentication < CompactIndexAPI
     end
   end
 end
+
+require_relative "helpers/artifice"
 
 Artifice.activate_with(CompactIndexStrictBasicAuthentication)
