+++ bundler/spec/support/artifice/compact_index_wrong_dependencies.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "compact_index"
-
-Artifice.deactivate
+require_relative "helpers/compact_index"
 
 class CompactIndexWrongDependencies < CompactIndexAPI
   get "/info/:name" do
@@ -13,5 +11,7 @@ class CompactIndexWrongDependencies < CompactIndexAPI
     end
   end
 end
+
+require_relative "helpers/artifice"
 
 Artifice.activate_with(CompactIndexWrongDependencies)
