+++ bundler/spec/support/artifice/endpoint_api_forbidden.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
-require_relative "endpoint"
-
-Artifice.deactivate
+require_relative "helpers/endpoint"
 
 class EndpointApiForbidden < Endpoint
   get "/api/v1/dependencies" do
     halt 403
   end
 end
+
+require_relative "helpers/artifice"
 
 Artifice.activate_with(EndpointApiForbidden)
