+++ bundler/spec/support/artifice/endpoint_basic_authentication.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "endpoint"
-
-Artifice.deactivate
+require_relative "helpers/endpoint"
 
 class EndpointBasicAuthentication < Endpoint
   before do
@@ -11,5 +9,7 @@ class EndpointBasicAuthentication < Endpoint
     end
   end
 end
+
+require_relative "helpers/artifice"
 
 Artifice.activate_with(EndpointBasicAuthentication)
