+++ bundler/spec/support/artifice/endpoint_host_redirect.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "endpoint"
-
-Artifice.deactivate
+require_relative "helpers/endpoint"
 
 class EndpointHostRedirect < Endpoint
   get "/fetch/actual/gem/:id", :host_name => "localgemserver.test" do
@@ -13,5 +11,7 @@ class EndpointHostRedirect < Endpoint
     status 404
   end
 end
+
+require_relative "helpers/artifice"
 
 Artifice.activate_with(EndpointHostRedirect)
