+++ bundler/spec/support/artifice/endpoint_marshal_fail.rb
@@ -1,13 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "endpoint_fallback"
-
-Artifice.deactivate
-
-class EndpointMarshalFail < EndpointFallback
-  get "/api/v1/dependencies" do
-    "f0283y01hasf"
-  end
-end
+require_relative "helpers/endpoint_marshal_fail"
+require_relative "helpers/artifice"
 
 Artifice.activate_with(EndpointMarshalFail)
