+++ bundler/spec/support/artifice/endpoint_strict_basic_authentication.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "endpoint"
-
-Artifice.deactivate
+require_relative "helpers/endpoint"
 
 class EndpointStrictBasicAuthentication < Endpoint
   before do
@@ -16,5 +14,7 @@ class EndpointStrictBasicAuthentication < Endpoint
     end
   end
 end
+
+require_relative "helpers/artifice"
 
 Artifice.activate_with(EndpointStrictBasicAuthentication)
