+++ bundler/spec/support/artifice/helpers/endpoint_extra.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "helpers/endpoint"
+require_relative "endpoint"
 
 class EndpointExtra < Endpoint
   get "/extra/api/v1/dependencies" do
@@ -27,7 +27,3 @@ class EndpointExtra < Endpoint
     File.binread("#{gem_repo2}/gems/#{params[:id]}")
   end
 end
-
-require_relative "helpers/artifice"
-
-Artifice.activate_with(EndpointExtra)
