+++ bundler/spec/support/artifice/windows.rb
@@ -4,10 +4,7 @@ require_relative "../path"
 
 $LOAD_PATH.unshift(*Dir[Spec::Path.base_system_gem_path.join("gems/{mustermann,rack,tilt,sinatra,ruby2_keywords}-*/lib")].map(&:to_s))
 
-require_relative "base"
 require "sinatra/base"
-
-Artifice.deactivate
 
 class Windows < Sinatra::Base
   set :raise_errors, true
@@ -42,5 +39,7 @@ class Windows < Sinatra::Base
     halt 500
   end
 end
+
+require_relative "helpers/artifice"
 
 Artifice.activate_with(Windows)
